% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanEst.R
\name{A0_estimate}
\alias{A0_estimate}
\title{Estimate A0}
\usage{
A0_estimate(A0data, variant = c("withLatent", "noLatent", "flatA0",
  "flatBoth"), cores = getOption("mc.cores", default =
  parallel::detectCores()), chains = 3L, iter = 2000L,
  method = c("sampling", "optimizing"), ...)
}
\arguments{
\item{A0data}{A list of data required, including known parameters}

\item{variant}{Which A0 variant to use: withLatent, noLatent, flatA0, flatBoth}

\item{cores}{Number of processing cores for running chains in parallel. 
See \code{?rstan::sampling}. Defaults to \code{parallel::detectCores}.}

\item{chains}{A positive integer specifying the number of Markov chains. 
The default is 3.}

\item{iter}{Number of iterations per chain (including warmup). Defaults to 2000.}

\item{method}{Either "sampling" (default) which runs the MC sampler, or "optimizing" 
which optimizes to find the maximum a posteriori estimate.}

\item{...}{Other arguments passed to rstan::sampling() for customizing the 
Monte Carlo sampler}
}
\description{
Fits model for A0 of one of several variants using Hamiltonian Monte Carlo.
}
