\name{A0est-package}
\alias{A0est-package}
\alias{A0est}
\docType{package}
\title{
\packageTitle{A0est}
}
\description{
\packageDescription{A0est}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{A0est}
\packageIndices{A0est}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{A0est}

Maintainer: \packageMaintainer{A0est}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
